tests/wpt/tests/scroll-animations/animation-trigger/parsing/timeline-trigger-exit-range-start-valid.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-animations-2/#animation-trigger">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
// https://drafts.csswg.org/scroll-animations-1/#view-timelines-ranges
test_valid_value("timeline-trigger-exit-range-start", "auto");
test_valid_value("timeline-trigger-exit-range-start", "normal");
test_valid_value("timeline-trigger-exit-range-start", "cover 0%", "cover");
test_valid_value("timeline-trigger-exit-range-start", "cover 100%");
test_valid_value("timeline-trigger-exit-range-start", "cover 120%");
test_valid_value("timeline-trigger-exit-range-start", "cover 42%");
test_valid_value("timeline-trigger-exit-range-start", "0", "0px");
test_valid_value("timeline-trigger-exit-range-start", "120%");
test_valid_value("timeline-trigger-exit-range-start", "120px");
test_valid_value("timeline-trigger-exit-range-start", "cover -42%");
test_valid_value("timeline-trigger-exit-range-start", "contain 42%");
test_valid_value("timeline-trigger-exit-range-start", "exit 42%");
test_valid_value("timeline-trigger-exit-range-start", "exit 1%, cover 2%, contain 0%", "exit 1%, cover 2%, contain");
test_valid_value("timeline-trigger-exit-range-start", "exit 1%, cover 2%, contain 100%");
test_valid_value("timeline-trigger-exit-range-start", "exit-crossing 42%");
test_valid_value("timeline-trigger-exit-range-start", "entry 42px");
test_valid_value("timeline-trigger-exit-range-start", "entry-crossing 42px");
test_valid_value("timeline-trigger-exit-range-start", "contain calc(10px + 10%)", "contain calc(10% + 10px)");
test_valid_value("timeline-trigger-exit-range-start", "entry 1em");
test_valid_value("timeline-trigger-exit-range-start", "exit calc(1em + 10px)");
test_valid_value("timeline-trigger-exit-range-start", "entry 42%");
test_valid_value("timeline-trigger-exit-range-start", "cover");
test_valid_value("timeline-trigger-exit-range-start", "contain");
test_valid_value("timeline-trigger-exit-range-start", "entry");
test_valid_value("timeline-trigger-exit-range-start", "exit");
test_valid_value("timeline-trigger-exit-range-start", "cover calc(sign(100em - 1px) * 1%)", "cover calc(1% * sign(100em - 1px))");
</script>
